requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==5.4.0
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==46.0.3
dnspython==2.6.1
et-xmlfile==1.1.0
fastapi==0.128.0
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
h11==0.16.0
httplib2==0.22.0
idna==3.7
joblib==1.4.2
nltk==3.8.1
obs-websocket-py==1.0
openpyxl==3.1.5
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.8.0
pyparsing==3.1.2
python-dotenv==1.2.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
rsa==4.9
schedule==1.2.2
starlette==0.49.3
tqdm==4.66.5
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.39.0
websocket-client==1.9.0
websockets==12.0
